Add backslash to URIPARAM
So I had this URL with \u00ab pattern, like "http://f00.ch/?q=Som\u0065+on\u0065"

The corresponding RFC does not explicit mention that backslashes are allowed in the query part, but nor denies it. To me it's like the backslash was forgotten when the spec was written.