Added Redis Monitor Log format
A common pattern is to pipe redis monitor output to a logfile for auditing or debugging purposes. 
The format of these logs though is a bit different, and to make it more confusing the command logged impacts the number of parameters. 
So as a simpler way to handle this I extract as much info as possible without getting into the intricacies of the the commands themselves: 

This was the data i tested against (with some manufactured pieces to cater for unknown future variations)

```
1470637867.953466 [0 195.168.1.1:52500] "info"
1339518083.107412 [0 127.0.0.1:60866] "keys" "*"
1470637875.777457 [15 195.168.1.1:52500] "select" "16"
1470637925.186681 [0 127.0.0.1:39404] "rpush" "my:special:key" "{\"data\":\"cdr\",\"payload\":\"json\"}"
1470637875.777457 [15 195.168.1.1:52500] "intentionally" "broken" "variadic" "log" "entry"
```

The params are simply not parsed, the more I tried to parse it the more it became apparent that you would have to handle it specifically per command, or at least for categories of commands.