Fix grokparsefailure due to truncated http_request.
The HAProxy documentation lists for httplog's http_request:

"This field might be truncated if the request is huge and does not fit
in the standard syslog buffer (1024 characters). This is the reason why
this field must always remain the last one."

If the field is truncated the ending quote will never be present and
thus fail the preceding regular expressions. By making the ending quote
optional we fill the http_request field, and http_version will be null.